requirements.txt
altgraph==0.17.4
black==23.10.1
click==8.1.7
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
pyserial==3.5
PySimpleGUI==4.60.5
python-dotenv==1.0.0
setuptools==70.0.0
numpy==1.26.4
